# docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "TaxiCity API",
		"description": "Comprehensive REST API for the TaxiCity taxi booking platform",
		"version": "1.0.0",
		"contact": {
			"name": "TaxiCity Support",
			"email": "support@taxiciti.com"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/api",
			"description": "Development server"
		},
		{
			"url": "https://taxiciti.vercel.app/api",
			"description": "Production server"
		}
	],
	"security": [
		{
			"ClerkAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"ClerkAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Clerk JWT authentication token"
			}
		},
		"schemas": {
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"clerkId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"phoneNumber": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": ["USER", "DRIVER", "ADMIN", "SUPPORT"]
					},
					"status": {
						"type": "string",
						"enum": ["ACTIVE", "INACTIVE", "SUSPENDED"]
					},
					"profilePicture": {
						"type": "string",
						"format": "uri"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"UserLocation": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"latitude": {
						"type": "number",
						"minimum": -90,
						"maximum": 90
					},
					"longitude": {
						"type": "number",
						"minimum": -180,
						"maximum": 180
					},
					"address": {
						"type": "string"
					},
					"accuracy": {
						"type": "number"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"Taxi": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"registrationNumber": {
						"type": "string"
					},
					"make": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"year": {
						"type": "integer"
					},
					"color": {
						"type": "string"
					},
					"capacity": {
						"type": "integer"
					},
					"status": {
						"type": "string",
						"enum": ["AVAILABLE", "BUSY", "OFFLINE", "MAINTENANCE"]
					},
					"currentRouteId": {
						"type": "string",
						"format": "uuid"
					},
					"currentRankId": {
						"type": "string",
						"format": "uuid"
					},
					"driverId": {
						"type": "string",
						"format": "uuid"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"Error": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string"
					},
					"details": {
						"type": "array",
						"items": {
							"type": "object"
						}
					}
				}
			},
			"PaginationMeta": {
				"type": "object",
				"properties": {
					"page": {
						"type": "integer"
					},
					"limit": {
						"type": "integer"
					},
					"total": {
						"type": "integer"
					},
					"totalPages": {
						"type": "integer"
					}
				}
			}
		}
	},
	"paths": {
		"/users": {
			"get": {
				"tags": ["Users"],
				"summary": "Get users list (admin only)",
				"description": "Retrieve a paginated list of all users. Only accessible by admin users.",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of users",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"users": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/User"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/PaginationMeta"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - Admin access required"
					}
				}
			},
			"post": {
				"tags": ["Users"],
				"summary": "Create or update user profile",
				"description": "Create a new user profile or update existing profile information",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["firstName", "lastName", "email"],
								"properties": {
									"firstName": {
										"type": "string",
										"minLength": 2,
										"maxLength": 50
									},
									"lastName": {
										"type": "string",
										"minLength": 2,
										"maxLength": 50
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"phoneNumber": {
										"type": "string"
									},
									"profilePicture": {
										"type": "string",
										"format": "uri"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User profile updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"201": {
						"description": "User profile created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input data"
					}
				}
			}
		},
		"/users/location": {
			"get": {
				"tags": ["Users"],
				"summary": "Get user's current location",
				"responses": {
					"200": {
						"description": "User's current location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLocation"
								}
							}
						}
					},
					"404": {
						"description": "Location not found"
					}
				}
			},
			"put": {
				"tags": ["Users"],
				"summary": "Update user's location",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["latitude", "longitude"],
								"properties": {
									"latitude": {
										"type": "number",
										"minimum": -90,
										"maximum": 90
									},
									"longitude": {
										"type": "number",
										"minimum": -180,
										"maximum": 180
									},
									"address": {
										"type": "string"
									},
									"accuracy": {
										"type": "number"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Location updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLocation"
								}
							}
						}
					}
				}
			}
		},
...
		"/taxis": {
			"get": {
				"tags": ["Taxis"],
				"summary": "Get taxis list",
				"parameters": [
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["AVAILABLE", "BUSY", "OFFLINE", "MAINTENANCE"]
						}
					},
					{
						"name": "routeId",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "rankId",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of taxis",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"taxis": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Taxi"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/PaginationMeta"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Taxis"],
				"summary": "Create a new taxi (admin only)",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["registrationNumber", "make", "model", "year", "capacity"],
								"properties": {
									"registrationNumber": {
										"type": "string"
									},
									"make": {
										"type": "string"
									},
									"model": {
										"type": "string"
									},
									"year": {
										"type": "integer",
										"minimum": 1990,
										"maximum": 2030
									},
									"color": {
										"type": "string"
									},
									"capacity": {
										"type": "integer",
										"minimum": 1,
										"maximum": 20
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Taxi created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Taxi"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}